tp2/requirements.txt
aiohttp
Pillow
logging
